+++ src/main/java/callprotector/spring/domain/abuse/dto/response/AbuseResponseDTO.java
@@ -1,4 +1,4 @@
-package callprotector.spring.web.dto.response;
+package callprotector.spring.domain.abuse.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
