+++ src/main/java/callprotector/spring/domain/abuse/entity/AbuseType.java
@@ -1,7 +1,6 @@
-package callprotector.spring.domain;
+package callprotector.spring.domain.abuse.entity;
 
 import callprotector.spring.global.common.entity.BaseEntity;
-import callprotector.spring.domain.mapping.AbuseTypeLog;
 import jakarta.persistence.*;
 import lombok.*;
