+++ src/main/java/callprotector/spring/domain/abuse/repository/AbuseTypeRepository.java
@@ -1,6 +1,6 @@
-package callprotector.spring.repository;
+package callprotector.spring.domain.abuse.repository;
 
-import callprotector.spring.domain.AbuseType;
+import callprotector.spring.domain.abuse.entity.AbuseType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AbuseTypeRepository extends JpaRepository<AbuseType, Long> {}
